django_backend/requirements.txt
Django>=4.2
psycopg2-binary>=2.9
djangorestframework
django-cors-headers
djangorestframework-simplejwt
Pillow
